--- a/cmd/ksuid/main.go
+++ b/cmd/ksuid/main.go
@@ -128,7 +128,7 @@ func printTemplate(id ksuid.KSUID) {
 		String    string
 		Raw       string
 		Time      time.Time
-		Timestamp uint32
+		Timestamp uint64
 		Payload   string
 	}{
 		String:    id.String(),