--- a/cmd/ksuid/main.go
+++ b/cmd/ksuid/main.go
@@ -11,7 +11,7 @@ import (
 	"text/template"
 	"time"
 
-	"github.com/segmentio/ksuid"
+	"github.com/signoz/ksuid"
 )
 
 var (