--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,3 @@
-module github.com/segmentio/ksuid
+module github.com/signoz/ksuid
 
 go 1.12