--- a/ksuid_test.go
+++ b/ksuid_test.go
@@ -308,8 +308,8 @@ func TestGetTimestamp(t *testing.T) {
 	nowTime := time.Now()
 	x, _ := NewRandomWithTime(nowTime)
 	xTime := int64(x.Timestamp())
-	unix := nowTime.Unix()
-	if xTime != unix - epochStamp {
+	unix := nowTime.UnixNano()
+	if xTime != unix-epochStamp {
 		t.Fatal(xTime, "!=", unix)
 	}
 }