Original 
* timestamp in seconds - 4 bytes
* random payload of 16 bytes

Updated
* timestamp in nanoseconds - 8 bytes
* random payload of 12 bytes

Wen't ahead with this approach as I don't want to increase the size of the ksuid.
The randomness won't be compromised because if you see the `BenchmarkAppendEncodeBase62` it's taking about a few ns. Even though we are compormising a bit on the random payload, the timestamp in ns will take care of it as `BenchmarkAppendEncodeBase62` doesn take few hundred ns for a operation. So it should be fine for our use case. 

Prev
```
➜  ksuid git:(master) ✗ go run cmd/ksuid/main.go -n 2                                  
2phFc2dzLsM9Zt1petCin8gfSSA
2phFc37W7uwadId97wQfkxu2yIu

➜  ksuid git:(master) ✗ go run cmd/ksuid/main.go -f inspect 2phFc2dzLsM9Zt1petCin8gfSSA                           

REPRESENTATION:

  String: 2phFc2dzLsM9Zt1petCin8gfSSA
     Raw: 13DC636BDD91C3F9473C1308C5B96E06912CF2DA

COMPONENTS:

       Time: 2024-12-03 12:51:15 +0530 IST
  Timestamp: 333210475
    Payload: DD91C3F9473C1308C5B96E06912CF2DA

➜  ksuid git:(master) ✗ go run cmd/ksuid/main.go -f inspect 2phFc37W7uwadId97wQfkxu2yIu                           

REPRESENTATION:

  String: 2phFc37W7uwadId97wQfkxu2yIu
     Raw: 13DC636BED37D384BC3E10A60C76EB2DD391D894

COMPONENTS:

       Time: 2024-12-03 12:51:15 +0530 IST
  Timestamp: 333210475
    Payload: ED37D384BC3E10A60C76EB2DD391D894

```


New

```
➜  ksuid git:(feat/epoch_nano_27_byte) ✗ go run cmd/ksuid/main.go -n 2                                  
0euHOuHoY2okD1KHtAImPc0jS7W
0euHOuHoZm2rD2OLvC5342XLWp0


➜  ksuid git:(feat/epoch_nano_27_byte) ✗ go run cmd/ksuid/main.go -f inspect 0euHOuHoY2okD1KHtAImPc0jS7W                           

REPRESENTATION:

  String: 0euHOuHoY2okD1KHtAImPc0jS7W
     Raw: 049FCCE76CC96E30A366D0885137AA3FA5D5AF5A

COMPONENTS:

       Time: 2024-12-03 12:46:31.783358 +0530 IST
  Timestamp: 333210191783358000
    Payload: A366D0885137AA3FA5D5AF5A

➜  ksuid git:(feat/epoch_nano_27_byte) ✗ go run cmd/ksuid/main.go -f inspect 0euHOuHoZm2rD2OLvC5342XLWp0                           

REPRESENTATION:

  String: 0euHOuHoZm2rD2OLvC5342XLWp0
     Raw: 049FCCE76CC97DD0E056AC26E414F1804553A582

COMPONENTS:

       Time: 2024-12-03 12:46:31.783362 +0530 IST
  Timestamp: 333210191783362000
    Payload: E056AC26E414F1804553A582
```



----
bench - original

```
goos: darwin
goarch: arm64
pkg: github.com/segmentio/ksuid
cpu: Apple M3 Pro
BenchmarkAppendEncodeBase62-11           3818517               597.4 ns/op
BenchmarkAppendFastEncodeBase62-11      15242377               157.6 ns/op
BenchmarkAppendDecodeBase62-11           3474770               704.5 ns/op
BenchmarkAppendFastDecodeBase62-11      29399379                84.05 ns/op
BenchmarkAppend-11                      15969307               150.7 ns/op
BenchmarkString-11                      14513085               176.0 ns/op
BenchmarkParse-11                       27735938                87.45 ns/op
BenchmarkCompare-11                     1000000000               1.620 ns/op
BenchmarkSort-11                         1000000              2155 ns/op
BenchmarkNew/with_crypto_rand-11         9551492               247.3 ns/op
BenchmarkNew/with_math_rand-11          58983396                39.99 ns/op
BenchmarkCompressedSet/write-11            56374             42925 ns/op         396.43 MB/s
BenchmarkCompressedSet/read-11            245125              9735 ns/op        1647.32 MB/s
BenchmarkCmp128-11                      1000000000               0.2508 ns/op
BenchmarkAdd128-11                      1000000000               0.2632 ns/op
BenchmarkSub128-11                      1000000000               0.2642 ns/op
PASS
ok      github.com/segmentio/ksuid      35.658s

```


bench - updated

```
goos: darwin
goarch: arm64
pkg: github.com/segmentio/ksuid
cpu: Apple M3 Pro
BenchmarkAppendEncodeBase62-11           3887671               572.1 ns/op
BenchmarkAppendFastEncodeBase62-11      18820423               128.8 ns/op
BenchmarkAppendDecodeBase62-11           3420231               717.3 ns/op
BenchmarkAppendFastDecodeBase62-11      30072784                83.15 ns/op
BenchmarkAppend-11                      17468864               137.1 ns/op
BenchmarkString-11                      16025907               152.0 ns/op
BenchmarkParse-11                       27031377                89.94 ns/op
BenchmarkCompare-11                     1000000000               1.680 ns/op
BenchmarkSort-11                          286790              7578 ns/op
BenchmarkNew/with_crypto_rand-11         9543036               251.2 ns/op
BenchmarkNew/with_math_rand-11          58344829                41.02 ns/op
BenchmarkCompressedSet/write-11            10000            222385 ns/op          62.95 MB/s
BenchmarkCompressedSet/read-11            330866              7163 ns/op        1817.65 MB/s
BenchmarkCmp128-11                      1000000000               0.2549 ns/op
BenchmarkAdd128-11                      1000000000               0.2554 ns/op
BenchmarkSub128-11                      1000000000               0.2646 ns/op
BenchmarkCmp96-11                       1000000000               0.2501 ns/op
BenchmarkAdd96-11                       1000000000               0.2565 ns/op
BenchmarkSub96-11                       1000000000               0.2505 ns/op
BenchmarkIncr96-11                      1000000000               0.2552 ns/op
PASS
ok      github.com/segmentio/ksuid      34.946s
```