@nityanandagohain 2024-12-05
I have updated the description with more details and a benchamark to test the collision, please have a look