@srikanthccv 2024-12-06
I reviewed it partially, I need some time to carefully follow the bit manipulation logic.