@srikanthccv 2024-12-07
The to and fro conversion b/w high 32 bits is confusing to follow. How about `[3]uint32` layout?