@nityanandagohain 2024-12-10
Updated it to use [3]uint32.